Add breadcrumbs back for Enumerated values
As enumerated values are not declared as subclasses of any
given type, our logic was skipping the creation of the
breadcrumb type hierarchy. Fall back to checking for typeOf
and creating a hierarchy that way.

In addition, we add a new "isEnumerationValue" method to
support checking whether a given type is a subclass of
Enumeration so that we can use the " :: " separator in
the breadcrumb trail.

isEnumerationValue() iterates over all of the arcs to look
for the signs of being an enumeration value (typeOf != rdfs:Class)
instead of assuming that the first incoming arc will be a
class instead of a property reference.

Fixes #10

Signed-off-by: Dan Scott <dan@coffeecode.net>